# API-SPEC.json
{
    "openapi":"3.0.2",
    "info": {
      "title":"BLOG API",
      "version":"1.0"
    },
    "servers": [
      {"url":"http://127.0.0.1:3001/api"}
    ],
    "paths": {
        "/users": {
            "get": {
              "security": [{
                "bearerAuth":[]
              }],
              "tags": ["blog API"],
              "description": "List all user",
              "summary": "List All user",
              "parameters": [],
              "responses": {
                "200": {
                  "description": "Success get all Customers",
                  "content": {
                    "application/json":{
                      "schema":{
                        "type": "object",
                        "properties": {
                          "code":{
                            "type": "number",
                            "example":200
                          },
                          "status":{
                            "type": "string",
                            "example":"OK"
                          },
                          "data":{
                            "type": "array",
                            "example":[
                              {
                                "id":"47568467394683",
                                "name":"Aditya",
                                "user_name":"adit",
                                "email":"adit@mail.com",
                                "created_at":"2022-11-01T10:23:47.362Z"
                              },
                              {
                                "id":"47568467394683",
                                "name":"Aditya",
                                "user_name":"adit",
                                "email":"adit@mail.com",
                                "created_at":"2022-11-01T10:23:47.362Z"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                },
                "500":{
                  "description": "Internal server error",
                  "content": {
                    "application/json":{
                      "schema":{
                        "type": "object",
                        "properties": {
                          "code":{
                            "example":500,
                            "type": "integer",
                            "format": "number"
                          },
                          "status":{
                            "example":"INTERNAL_SERVER_ERROR",
                            "type": "string",
                            "format": "string"
                          },
                          "data":{
                            "type": "object",
                            "example":null
                          }
                        }
                      }
                    }
                  }
                } 
              }
            },
            "post":{
              "security": [{
                "bearerAuth":[]
              }],
              "tags": ["blog API"],
              "description": "Create User",
              "summary": "Create User",
              "parameters": [],
              "requestBody":{
                "content": {
                "application/json":{
                  "schema":{
                    "type": "object",
                    "properties": {
                      "user_name":{
                        "type": "string",
                        "example":"adit",
                        "description": "Username user"
                      },
                      "name":{
                        "type": "string",
                        "example":"Aditya",
                        "description": "Nama User" 
                      },
                      "email":{
                        "type": "string",
                        "example":"adit@mail.com",
                        "description": "Email User"
                      },
                      "password":{
                        "type": "string",
                        "example":"admin",
                        "description": "Password User"
                      },
                      "address":{
                        "type": "string",
                        "example":"Jl MOh Toha",
                        "description": "Alamat user"
                      }
                    },
                    "required": ["user_name","name","email"]
                  }
                }
                }
              },"responses": {
              "200":{
                "description": "Succes Register",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "example":200,
                          "type": "integer",
                          "format": "number"
                        },
                        "status":{
                          "example":"OK",
                          "type": "string",
                          "format": "string"
                        },
                        "data":{
                          "type": "object",
                          "example":{
                            "id":"47568467394683",
                            "name":"Aditya",
                            "user_name":"adit",
                            "email":"adit@mail.com",
                            "created_at":"2022-11-01T10:23:47.362Z"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "400":{
                "description": "Validasi",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "integer",
                          "example":400
                        },
                        "status":{
                          "type": "string",
                          "example":"BAD_REQUEST"
                        },
                        "data":{
                          "type": "array",
                          "example":[
                            {
                              "field":"UserName",
                              "message":"This field is required"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              },
              "500":{
                "description": "Internal server error",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "example":500,
                          "type": "integer",
                          "format": "number"
                        },
                        "status":{
                          "example":"INTERNAL_SERVER_ERROR",
                          "type": "string",
                          "format": "string"
                        },
                        "data":{
                          "type": "object",
                          "example":null
                        }
                      }
                    }
                  }
                }
              } 
              }
            }
        },
        "/users/{id}":{
          "get":{
            "security": [{
              "bearerAuth":[]
            }],
            "tags": ["blog API"],
            "description": "Get user by id",
            "summary": "Get user by id",
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "User Id"
              }
            ],
            "responses": {
              "200": {
                "description": "Success get user",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "number",
                          "example":200
                        },
                        "status":{
                          "type": "string",
                          "example":"OK"
                        },
                        "data":{
                          "type": "object",
                          "example":
                            {
                              "id":"47568467394683",
                              "name":"Aditya",
                              "user_name":"adit",
                              "email":"adit@mail.com",
                              "created_at":"2022-11-01T10:23:47.362Z"
                            }
                        }
                      }
                    }
                  }
                }
              },
              "404": {
                "description": "User Not Found",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "number",
                          "example":404
                        },
                        "status":{
                          "type": "string",
                          "example":"NOT_FOUND"
                        },
                        "data":{
                          "type": "object",
                          "example":null        
                        }
                      }
                    }
                  }
                }
              },
              "500":{
                "description": "Internal server error",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "example":500,
                          "type": "integer",
                          "format": "number"
                        },
                        "status":{
                          "example":"INTERNAL_SERVER_ERROR",
                          "type": "string",
                          "format": "string"
                        },
                        "data":{
                          "type": "object",
                          "example":null
                        }
                      }
                    }
                  }
                }
              } 
            }
          },
          "put":{
            "security": [{
              "CategoryAuth":[]
            }],
            "tags": ["blog API"],
            "description": "Edit user by id",
            "summary": "Edit user by id",
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "User Id"
              }
            ],
            "requestBody":{
              "content": {
              "application/json":{
                "schema":{
                  "type": "object",
                  "properties": {
                    "name":{
                      "type": "string",
                      "example":"Aditya",
                      "description": "Nama User" 
                    },
                    "address":{
                      "type": "string",
                      "example":"Jl MOh Toha",
                      "description": "Alamat user"
                    }
                  },
                  "required": ["name"]
                }
              }
              }
            },
            "responses": {
              "200": {
                "description": "Success Edit user",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "number",
                          "example":200
                        },
                        "status":{
                          "type": "string",
                          "example":"OK"
                        },
                        "data":{
                          "type": "object",
                          "example":
                            {
                              "id":"47568467394683",
                              "name":"Aditya",
                              "user_name":"adit",
                              "email":"adit@mail.com",
                              "created_at":"2022-11-01T10:23:47.362Z"
                            }
                        }
                      }
                    }
                  }
                }
              },
              "400":{
                "description": "Validasi",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "integer",
                          "example":400
                        },
                        "status":{
                          "type": "string",
                          "example":"BAD_REQUEST"
                        },
                        "data":{
                          "type": "array",
                          "example":[
                            {
                              "field":"UserName",
                              "message":"This field is required"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              },
              "404": {
                "description": "User Not Found",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "number",
                          "example":404
                        },
                        "status":{
                          "type": "string",
                          "example":"NOT_FOUND"
                        },
                        "data":{
                          "type": "object",
                          "example":null        
                        }
                      }
                    }
                  }
                }
              },
              "500":{
                "description": "Internal server error",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "example":500,
                          "type": "integer",
                          "format": "number"
                        },
                        "status":{
                          "example":"INTERNAL_SERVER_ERROR",
                          "type": "string",
                          "format": "string"
                        },
                        "data":{
                          "type": "object",
                          "example":null
                        }
                      }
                    }
                  }
                }
              } 
            }
          },
          "delete":{
            "security": [{
              "bearerAuth":[]
            }],
            "tags": ["blog API"],
            "description": "Delete user by id",
            "summary": "Delete user by id",
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "User Id"
              }
            ],
            "responses": {
              "200": {
                "description": "Success Delete user",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "number",
                          "example":200
                        },
                        "status":{
                          "type": "string",
                          "example":"OK"
                        },
                        "data":{
                          "type": "object",
                          "example":
                            {
                              "id":"47568467394683",
                              "name":"Aditya",
                              "user_name":"adit",
                              "email":"adit@mail.com",
                              "created_at":"2022-11-01T10:23:47.362Z"
                            }
                        }
                      }
                    }
                  }
                }
              },
              "404": {
                "description": "User Not Found",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "type": "number",
                          "example":404
                        },
                        "status":{
                          "type": "string",
                          "example":"NOT_FOUND"
                        },
                        "data":{
                          "type": "object",
                          "example":null        
                        }
                      }
                    }
                  }
                }
              },
              "500":{
                "description": "Internal server error",
                "content": {
                  "application/json":{
                    "schema":{
                      "type": "object",
                      "properties": {
                        "code":{
                          "example":500,
                          "type": "integer",
                          "format": "number"
                        },
                        "status":{
                          "example":"INTERNAL_SERVER_ERROR",
                          "type": "string",
                          "format": "string"
                        },
                        "data":{
                          "type": "object",
                          "example":null
                        }
                      }
                    }
                  }
                }
              } 
            }
          }
        },
        "/login":{
          "post":{
            "security": [{
              "bearerAuth":[]
            }],
            "tags": ["blog API"],
            "description": "Login User",
            "summary": "Login User",
            "parameters": [],
            "requestBody":{
              "content": {
              "application/json":{
                "schema":{
                  "type": "object",
                  "properties": {
                    "email":{
                      "type": "string",
                      "example":"admin@mail.com",
                      "description": "Email User"
                    },
                    "password":{
                      "type": "string",
                      "example":"admin",
                      "description": "Password User"
                    }
                  },
                  "required": ["email","password"]
                }
              }
              }
            },"responses": {
            "200":{
              "description": "Succes Login",
              "content": {
                "application/json":{
                  "schema":{
                    "type": "object",
                    "properties": {
                      "code":{
                        "example":200,
                        "type": "integer",
                        "format": "number"
                      },
                      "status":{
                        "example":"OK",
                        "type": "string",
                        "format": "string"
                      },
                      "data":{
                        "type": "object",
                        "example":{
                          "id":"47568467394683",
                          "name":"Aditya",
                          "user_name":"adit",
                          "email":"adit@mail.com",
                          "created_at":"2022-11-01T10:23:47.362Z"
                        }
                      }
                    }
                  }
                }
              }
            },
            "400":{
              "description": "Validasi",
              "content": {
                "application/json":{
                  "schema":{
                    "type": "object",
                    "properties": {
                      "code":{
                        "type": "integer",
                        "example":400
                      },
                      "status":{
                        "type": "string",
                        "example":"BAD_REQUEST"
                      },
                      "data":{
                        "type": "array",
                        "example":[
                          {
                            "field":"UserName",
                            "message":"This field is required"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "500":{
              "description": "Internal server error",
              "content": {
                "application/json":{
                  "schema":{
                    "type": "object",
                    "properties": {
                      "code":{
                        "example":500,
                        "type": "integer",
                        "format": "number"
                      },
                      "status":{
                        "example":"INTERNAL_SERVER_ERROR",
                        "type": "string",
                        "format": "string"
                      },
                      "data":{
                        "type": "object",
                        "example":null
                      }
                    }
                  }
                }
              }
            } 
            }
          }
        },
...
    },
    "components": {
      "securitySchemes": {
        "bearerAuth":{
          "type":"http",
          "scheme": "bearer",
          "in":"header",
          "bearerFormat": "JWT",
          "name":"Bearer-KEY",
          "description":"Authentication for API"
        }
      },
...
    }
  }